requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
langchain-openai
langgraph
